[metasrv] refactor: move node_id tests from sled_store/ into meta_service/
